Update README.md
minor updates to syntax. 
Also noted issue with docker-compose version 3.  Appears to be unsupported switching to version 2 worked as expected.  Error message was "ERROR: Version in "./docker-compose.yaml" is unsupported. You might be seeing this error because you're using the wrong Compose file version. Either specify a version of "2" (or "2.0") and place your service definitions under the `services` key, or omit the `version` key and place your service definitions at the root of the file to use version 1."